Refactor tuist setup
Split Project.swift into extensions to make it easier to create multiple targets in the future.
