--- Tuist/Tuist.swift
@@ -1,8 +1,0 @@
-import ProjectDescription
-
-let config = Config(
-    swiftVersion: "5.6.1",
-    plugins: [
-        .local(path: .relativeToManifest("../../Plugins/Tuist")),
-    ]
-)
